警告の改善&コメント追加

+++ Project/GameSources/SelectStage.cpp
@@ -73,7 +73,7 @@ namespace basecross {
 				position , Vec2(212.5f,335.0f) * 1.5f, 
 				[](shared_ptr<ObjectInterface> object) {
 					auto stage = static_pointer_cast<SelectStage>(object);
-					stage->AcceptStage(ButtonManager::instance->GetSelectIndex(L"City"));
+					stage->AcceptStage(static_cast<int>(ButtonManager::instance->GetSelectIndex(L"City")));
 				});
 
 			AddGameObject<Sprite>(L"SELECT_STAGE", position + Vec3(0.0f,175.0f,0.0f), Vec2(150.0f, 50.0f), true);
@@ -83,8 +83,8 @@ namespace basecross {
 					dangerLow - Vec3(90.0f, 100.0f * j, 0.0f), Vec2(60.0f, 60.0f),
 					[](shared_ptr<ObjectInterface> object) {
 						auto stage = static_pointer_cast<SelectStage>(object);
-						int selectCity = ButtonManager::instance->GetSelectIndex(L"City");
-						stage->AcceptDifficulty(ButtonManager::instance->GetSelectIndex(L"Difficulty" + to_wstring(selectCity)));
+						size_t selectCity = ButtonManager::instance->GetSelectIndex(L"City");
+						stage->AcceptDifficulty(static_cast<int>(ButtonManager::instance->GetSelectIndex(L"Difficulty" + to_wstring(selectCity))));
 					});
 
 				auto sprite = AddGameObject<Sprite>(dangerKey[j], dangerLow - Vec3(0,100.0f * j,0.0f), Vec2(60.0f, 60.0f), true);
